Feature : 회원삭제 스케줄링 추가 / DB조회 시 회원탈퇴 여부 확인 로직 추가(isDeletedUser)

+++ src/services/user.service.js
@@ -47,7 +47,11 @@ export const userProfile = async (userId) => {
 
 
 export const userDelete = async (userId, user) => {
+    const isUserExist = await getMyProfile(userId);
+    if (!isUserExist) throw new NoExistsUserError(userId);
+
     const deletedUser = await deleteUser(userId);
+
 
     console.log('DB에서 유저 삭제 성공');
